+++ examples/recurrence/lib/toolbar.dart
@@ -23,7 +23,7 @@ class CalendarToolBar extends StatelessWidget {
         spacing: 4,
         children: [
           ValueListenableBuilder(
-            valueListenable: kalenderController.floatingRange,
+            valueListenable: kalenderController.floatingVisibleRange,
             builder: (context, value, child) {
               if (value == null) return const SizedBox.shrink();
               final localRange = value.forLocation();